dev updated

+++ src/main/java/com/example/gym_management/repository/PaymentRepository.java
@@ -5,6 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
@@ -18,6 +19,6 @@ public interface PaymentRepository extends JpaRepository<Payment, Long> {
     List<Payment> findExpiredPaymentsOfActiveMembers(@Param("date") LocalDate date);
 
     @Query("SELECT SUM(p.amount) FROM Payment p WHERE p.paymentDate BETWEEN :start AND :end")
-    Optional<Double> sumAmountByPaymentDateBetween(@Param("start") LocalDate start, @Param("end") LocalDate end);
+    Optional<BigDecimal> sumAmountByPaymentDateBetween(@Param("start") LocalDate start, @Param("end") LocalDate end);
 
 }
